src/webapp/backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
google-cloud-bigquery==3.38.0
python-dotenv==1.2.1
pydantic==2.12.5
